backend/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.2
Pillow==8.1.1
pydantic-django==0.0.7
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.8.6